+++ frontend/src/pages/SharePage/SharePage.tsx
@@ -1,9 +1,8 @@
-import React, { useEffect } from 'react'
+import React, { useEffect, useState } from 'react'
 import { useParams, useLocation } from 'react-router-dom'
 import { useDispatch, useSelector } from 'react-redux'
 import { Dispatch, ApplicationState } from '../../store'
 import { makeStyles, Typography, IconButton, Tooltip, CircularProgress } from '@material-ui/core'
-import { DeviceShareContainer } from '../../components/DeviceShareContainer'
 import { Breadcrumbs } from '../../components/Breadcrumbs'
 import { findService } from '../../models/devices'
 import { Container } from '../../components/Container'
@@ -12,6 +11,8 @@ import { Icon } from '../../components/Icon'
 import { useHistory } from 'react-router-dom'
 import analyticsHelper from '../../helpers/analyticsHelper'
 import styles from '../../styling'
+import { DeviceShareAdd, DeviceShareDetails } from '../../components/DeviceShareContainer'
+import { SharingDetails } from '../../components/DeviceShareContainer/SharingForm'
 
 export const SharePage = () => {
   const { email = '', deviceID = '', serviceID = '' } = useParams()
@@ -27,6 +28,9 @@ export const SharePage = () => {
     }
     return { deleting, device }
   })
+  const { contacts = [] } = useSelector((state: ApplicationState) => state.devices)
+  const [selectedContacts, setSelectedContacts] = React.useState<string[]>([])
+  const [changing, setChanging] = useState(false)
   const location = useLocation()
   const history = useHistory()
   const css = useStyles()
@@ -40,12 +44,48 @@ export const SharePage = () => {
     history.push(location.pathname.replace(email ? `/${email}` : '/share', ''))
   }
 
+  const handleShareUpdate = (share: SharingDetails, isNew: boolean) => {
+    const shareData = mapShareData(share, isNew)
+    const { scripting, services } = share.access
+    shares.share(shareData)
+    if (device && shareData) {
+      shares.updateDeviceState({ device, contacts: shareData.email, scripting, services, isNew })
+    }
+    goToNext()
+  }
+
+  const goToNext = () =>
+    email === ''
+      ? history.push(location.pathname.replace('/share', ''))
+      : history.push(location.pathname.replace(`/${email}`, ''))
+
+  const mapShareData = (share: SharingDetails, isNew: boolean) => {
+    const { access } = share
+    const scripting = access.scripting
+    const services =
+      device?.services.map((ser: { id: string }) => ({
+        serviceId: ser.id,
+        action: access.services.includes(ser.id) ? 'ADD' : 'REMOVE',
+      })) || []
+    const email = isNew ? share.contacts : [share.contacts[0]]
+
+    if (device) {
+      return {
+        deviceId: device.id,
+        scripting,
+        services,
+        email,
+      }
+    }
+    return
+  }
+
   return (
     <Container
       header={
         <>
           <Breadcrumbs />
-          {email && (
+          {email ? (
             <Typography variant="h1">
               <Icon name={email === '' ? 'user-plus' : 'user'} size="lg" />
               <Title>{email || 'Share'}</Title>
@@ -59,11 +99,34 @@ export const SharePage = () => {
                 </Tooltip>
               )}
             </Typography>
+          ) : (
+            device && (
+              <DeviceShareAdd
+                contacts={contacts}
+                device={device}
+                onChangeContacts={setSelectedContacts}
+                selectedContacts={selectedContacts}
+                setChanging={setChanging}
+              />
+            )
           )}
         </>
       }
     >
-      <DeviceShareContainer device={device} email={email} />
+      {
+        <>
+          {device && (
+            <DeviceShareDetails
+              device={device}
+              share={handleShareUpdate}
+              selectedContacts={selectedContacts}
+              updateSharing={handleShareUpdate}
+              changing={changing}
+              setChanging={setChanging}
+            />
+          )}
+        </>
+      }
     </Container>
   )
 }
